PDF Export
Generate PDF for Dashboard tab.
